+++ routes/web.php
@@ -1,11 +1,18 @@
 <?php
 
+use App\Http\Controllers\AdminController;
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\CategorieController;
 use App\Http\Controllers\CustomerController;
 use App\Http\Controllers\DashboardController;
 use App\Http\Controllers\ProductController;
-use App\Http\Controllers\UserController;
+use App\Http\Controllers\SubCategoriController;
+use App\Http\Controllers\User\AuthController as UserAuthController;
+use App\Http\Controllers\User\HomepageController;
+use App\Http\Controllers\User\ProductController as UserProductController;
+use App\Http\Controllers\User\ProfileController;
+use App\Http\Controllers\User\SellController;
+use App\Http\Controllers\User\WistlistController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -19,25 +26,65 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::middleware('guest')->group(function() {
-    Route::get('/admin-registrasi', [AuthController::class, 'registrasi'])->name('registrasi');
-    Route::post('/admin-registrasi', [AuthController::class, 'store'])->name('auth.store');
-    Route::get('/admin-login', [AuthController::class, 'login'])->name('login');
-    Route::post('/admin-login', [AuthController::class, 'authenticate'])->name('auth.login');
+Route::middleware('isGuest')->group(function() {
+    // Route Login Admin
+    Route::get('/admin/login', [AuthController::class, 'login'])->name('adminLogin');
+    Route::post('/admin/login', [AuthController::class, 'authenticateAdmin'])->name('authAdminlogin');
+
+    //=================================================//
+
+    // Route Login Pelanggan
+    Route::get('/', [UserProductController::class, 'index'])->name('home');
+    Route::post('/auth-login', [UserAuthController::class, 'authenticateUser'])->name('authUserlogin');
+    Route::post('/auth-registrasi', [UserAuthController::class, 'registrasi'])->name('registrasi');
+
+    // route google
+    Route::get('/auth/redirect', [UserAuthController::class, 'redirectGoogle'])->name('redirectGoogle');
+    Route::get('/google/redirect', [UserAuthController::class, 'googleCallBack'])->name('googleCallBack');
+
 });
 
-
-Route::middleware('auth')->group(function(){
-    Route::get('/', [DashboardController::class, 'index'])->name('dashboard');
+Route::middleware('isAdmin')->group(function(){
+    Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
     Route::get('/categorie', [CategorieController::class, 'index'])->name('categorie');
     Route::post('/categorie/create', [CategorieController::class, 'store'])->name('categorie.store');
     Route::put('/categorie/{id}', [CategorieController::class, 'update'])->name('categorie.update');
     Route::delete('/categorie/{id}', [CategorieController::class, 'destroy'])->name('categorie.delete');
-    Route::get('product', [ProductController::class, 'index'])->name('product');
-    Route::put('/product/status/{id}', [ProductController::class, 'updateStatus'])->name('product.updateStatus');
-    Route::delete('/product/{id}', [ProductController::class, 'destroy'])->name('product.delete');
-    Route::get('customer', [CustomerController::class, 'index'])->name('customer');
-    Route::get('profile', [UserController::class, 'index'])->name('profile');
-    Route::put('/profile/update', [UserController::class, 'updateProfile'])->name('profile.update');
-    Route::post('/logout', [AuthController::class, 'logout'])->name('logout');
+    Route::get('/subcategorie', [SubCategoriController::class, 'index'])->name('subcategorie');
+    Route::post('/subcategorie/create', [SubCategoriController::class, 'store'])->name('subcategorie.store');
+    Route::put('/subcategorie/{id}', [SubCategoriController::class, 'update'])->name('subcategorie.update');
+    Route::delete('/subcategorie/{id}', [SubCategoriController::class, 'destroy'])->name('subcategorie.delete');
+    Route::get('/admin/product', [ProductController::class, 'index'])->name('product');
+    Route::put('/admin/product/status/{id}', [ProductController::class, 'updateStatus'])->name('product.updateStatus');
+    Route::delete('/admin/product/{id}', [ProductController::class, 'destroy'])->name('product.delete');
+    Route::get('/admin/user', [CustomerController::class, 'index'])->name('user');
+    Route::get('/admin', [AdminController::class, 'index'])->name('admin');
+    Route::post('/admin/user/create', [AdminController::class, 'store'])->name('admin.store');
+    Route::get('/admin/profile', [AdminController::class, 'myProfile'])->name('admin.profile');
+    Route::put('/admin/profile/update', [AdminController::class, 'updateProfile'])->name('admin.upProfile');
+    Route::put('/admin/profile/update/password', [AdminController::class, 'updatePassword'])->name('admin.upPassword');
+    Route::post('/admin/logout', [AuthController::class, 'logout'])->name('adminLogout')->middleware('onlyPost');
+});
+
+Route::middleware('isUser')->group(function(){
+    Route::get('/home', [UserProductController::class, 'index'])->name('homepage');
+    Route::post('/user/logout', [UserAuthController::class, 'logout'])->name('userLogout');
+    Route::get('/sell', [SellController::class, 'categorie'])->name('sell');
+    Route::post('/sell/create', [SellController::class, 'store'])->name('sell.store');
+    Route::get('/wishlist', [WistlistController::class, 'index'])->name('wistlist');
+    Route::delete('/wishlist/{id}', [WistlistController::class, 'destroy'])->name('wistlist.delete');
+    Route::post('/wishlist/toggle', [WistlistController::class, 'toggle'])->name('wishlist.toggle');
+    Route::get('/product',[UserProductController::class, 'product'])->name('web.product');
+    Route::get('/subcategories/{id}', [UserProductController::class, 'getSubcategories'])->name('web.subcategorie');
+    Route::get('/product/detail/{id}', [UserProductController::class, 'detail'])->name('web.product.detail');
+    Route::delete('/product/{id}', [UserProductController::class, 'destroy'])->name('web.product.destroy');
+    Route::get('/profile', [ProfileController::class, 'index'])->name('web.profile');
+    Route::get('/profile/edit', [ProfileController::class, 'edit'])->name('web.profile.edit');
+    Route::put('/profile/update', [ProfileController::class, 'update'])->name('web.profile.update');
+    Route::get('/user/seller/{id}', [ProfileController::class, 'seller'])->name('web.profile.seller');
+
+    Route::get('/nego', function(){
+        return view('website.nego.index');
+    });
+
 });
